move ddsFileSet to a different file
